+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - Add support for type checking scalar fields. This introduces a new warning
   message "protobuf-type-error"
 - Add warning for use of positional arguments on message constructors. This
-  introduces a new warning message "protobuf-type-error"
+  introduces a new warning message "protobuf-no-posargs"
+- Add support for type checking message constructor arguments
 
 ## [0.17.0] - 2021-02-18
 - Add support for scalar repeated fields